# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "iProjectIS2"
  description: Api's for the project.
host: iprojectis2.herokuapp.com
basePath: "/v1"
...
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /exams:
    x-summary: /exams
    get:
      tags: 
        - "Exam"
      summary: List of Exams
      produces:
      - "application/json"
      parameters:
      - in: "query"
        name: "idExams"
        type: "integer"
      - in: "query"
        name: "nameExams"
        type: "string"
      - in: "query"
        name: "idTest"
        type: "integer"
      - in: "query"
        name: "idProfessor"
        type: "integer"
      - in: "query"
        name: "nameProfessor"
        type: "string"
      - in: "query"
        name: "idCourse"
        type: "integer"
      - in: "query"
        name: "nameCourse"
        type: "string"
      - in: "query"
        name: "exam-type"
        type: "string"
        enum:
        - "oral"
        - "write"
        - "practical"
      responses:
        200:
          description: 'Successful Response'
          schema: 
            type: "array"
            items:
              $ref: "#/definitions/Exams"
        401:
          description: "invalid Input"
          
    post:
      summary: Create Exam
      produces: 
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Exam object that needs to be added"
        schema:
          $ref: "#/definitions/Exams"
         
      responses:
        201:
          description: 'Succesful'
        405:
          $ref: "#/responses/UnauthorizedError"
      security: 
      - APIKey_query: []
      - APIKey_header: []
    put:
      summary: Update Exams
      produces: 
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Update Exam"
        schema:
          $ref: "#/definitions/Exams"
      responses:
        201:
          description: 'Valid'
        400:
          description: "Invalid exam supplied"
        404:
          description: "Exam not found"
      security: 
      - APIKey_query: []
      - APIKey_header: []
  /exams/{idExam}:
    delete:
      tags: 
        - "Exam"
      summary: Delete Exam By Id
      operationId: "DeleteExamById"
      produces: 
      - "application/json"
      responses:
        204: 
          description: 'successful'
        400:
          description: 'Invalid ID'
        404: 
          description: 'not found'
      security: 
      - APIKey_query: []
      - APIKey_header: []
...
definitions:
...
  Exams: 
   type: "object"
   properties:
    idExam:
      type: "integer"
    nameExam: 
      type: "string"
    idProfessor:
      type: "integer"
    idCourse: 
      type: "integer"
    deadline:
      type: "string"
      format: "date-time"
    examType:
      type: "string"
      description: "Type of Exam"
      enum:
        - "oral"
        - "write"
        - "practical"
...
responses:
  UnauthorizedError:
    description: API key is missing or invalid
    headers: 
      WWW_Authenticate:
        type: string
securityDefinitions:
  APIKey_query:
     type: apiKey
     in: query
     name: token_in_query
  APIKey_header:
     type: apiKey
     in: header
     name: token_in_header
...